pyproject.toml
[project]
name = "TheRealUpstateNY"
version = "0.0.0"
description = "Bot that posts random maps of New York State with random counties highlighted indicating 'Upstate NY'"
authors = [
  { name = "Aaron Marriner", email = "amarriner@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "polyline (>=2.0.2,<3.0.0)",
    "atproto (>=0.0.57,<0.0.58)",
    "requests (>=2.32.3,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

